WIP No context.user_ptr for now
Seems I'm doing something wrong with it, that freezes my window manager
for some reasons, so I'll read more about it for next time.

The main draw of that was that you did not need to change much when you
use commodino and raph_raylib, but that will have to wait.
For now
g.input is passed explicitely.