BAD TRY double_buffer Session_Memory for frame replay
Seems I messed something up in the game when replacing the `g.` with
`g.old_s.` and `g.s.`

As I currently can move the player and have one sheep that jumps into
the pit, but no more sheeps while the lava does go down as-if sheeps
were sacrificed.

I'll retry later, or go directly to the "replay X frames from assert"
approach, using the record-and-replay checkpoints and the records to
loop over multiple frames and repeat.